+++ Packages/com.thenathannator.plasticband/CHANGELOG.md
@@ -10,7 +10,7 @@ Dates are relative to UTC.
 
 ### Added
 
-- Wii Rock Band guitars are now supported.
+- Wii Rock Band guitars and drumkits are now supported.
 - PS3 and Wii Pro Guitars are now supported.
 - PS4 Rock Band guitars are now supported.